moodibuddy/requirements.txt
asgiref==3.2.10
Django==3.1
django-debug-toolbar==2.2
djangorestframework==3.11.1
psycopg2-binary==2.8.6
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.3.1
uwsgi==2.0.19.1
django-cors-headers==3.6.0
